Make the BT turn ON/OFF have the same behavior as the button
To turn ON or OFF the Bluetooth in the settings, there is two
possibilities and they don't have the same behavior
- The Button: when pressed, blocks until a state (ON or OFF) is reached
- The SwitchBar (containing the Button): can be pressed whatever the
    state of the BT (ON, OFF or between)

When pressed, the SwitchBar set the state of the button which then calls
the switch function. The SwitchBAr does that wihout checking if the
Button is cliquable, hence changing the state the BT should go to.

Even if the implementation should handle it, when pressing repeatdly and
quickly the SwitchBar, at some point, the SwitchBar hung and the BT is
in an unknown state.
So the user has to leave the menu and comeback.

This patch replaces all calls of setting the Button state by calling
directly the switch function adn setting the state of the SwitchBar
accordingly.

This different behavior is only seen with BT SwitchBar. In effect, in
Wi-Fi settings, the SwitchBar has the same behavior than the button.

Change-Id: I032b3cbd1095df94f33a4d43c9dc5f1ea4053d19
Signed-off-by: JaganathX Kanakkassery <jaganathx.kanakkassery@intel.com>
Signed-off-by: Zhiquan Liu <zhiquan.liu@intel.com>